requirements.txt
Django==2.1.3
djangorestframework==3.9.0
names==0.3.0
pytz==2018.7
